extract outputs formatting to dedicated file
