+++ scopes/workspace/workspace-config-files/ws-config.cmd.ts
@@ -4,17 +4,18 @@ import { omit } from 'lodash';
 import { Command, CommandOptions } from '@teambit/cli';
 import chalk from 'chalk';
 import {
-  AspectWritersResults,
-  EnvsWrittenConfigFile,
-  EnvsWrittenConfigFiles,
-  EnvsWrittenExtendingConfigFile,
-  EnvsWrittenExtendingConfigFiles,
-  OneConfigFileWriterResult,
   WorkspaceConfigFilesMain,
-  WriteConfigFilesResult,
+  WriteConfigFilesResult
 } from './workspace-config-files.main.runtime';
+import { formatWriteOutput } from './format-outputs';
 
-type Flags = { dryRun?: boolean; noDedupe?: boolean; dryRunWithContent?: boolean; clean?: boolean; silent?: boolean };
+export type Flags = {
+  dryRun?: boolean;
+  noDedupe?: boolean;
+  dryRunWithContent?: boolean;
+  clean?: boolean;
+  silent?: boolean;
+};
 
 export class WsConfigCmd implements Command {
   name = 'ws-config <sub-command>';
@@ -57,11 +58,11 @@ export default class WsConfigWriteCmd implements Command {
   constructor(private workspaceConfigFilesMain: WorkspaceConfigFilesMain) {}
 
   async report(_args, flags: Flags) {
-    const results = await this.json(_args, flags) as WriteConfigFilesResult;
+    const results = (await this.json(_args, flags)) as WriteConfigFilesResult;
     if (flags.dryRunWithContent) {
       throw new Error(`use --json flag along with --dry-run-with-content`);
     }
-    return this.formatOutput(results, flags);
+    return formatWriteOutput(results, flags);
   }
 
   async json(_args, flags: Flags) {
@@ -86,75 +87,5 @@ export default class WsConfigWriteCmd implements Command {
       };
     }
     return { cleanResults, writeResults };
-  }
-
-  private formatOutput(writeConfigFilesResult: WriteConfigFilesResult, flags: Flags): string {
-    const { cleanResults, writeResults } = writeConfigFilesResult;
-    const isDryRun = !!(flags.dryRun || flags.dryRunWithContent);
-    const cleanResultsOutput = this.getCleanResultsOutput(cleanResults, isDryRun);
-    const writeResultsOutput = this.getWriteResultsOutput(writeResults, isDryRun);
-
-    return `${cleanResultsOutput}${writeResultsOutput}`;
-  }
-
-  private getCleanResultsOutput(cleanResults: string[] | undefined, isDryRun: boolean) {
-    const cleanResultsOutput = cleanResults
-      ? `${chalk.green(
-          `the following ${cleanResults.length} paths ${isDryRun ? 'will be' : 'were'} deleted`
-        )}\n${cleanResults.join('\n')}\n\n`
-      : '';
-    return cleanResultsOutput;
-  }
-  private getWriteResultsOutput(writeResults: WriteConfigFilesResult['writeResults'], isDryRun: boolean) {
-    const totalFiles = writeResults.totalWrittenFiles;
-
-    const writeTitle = isDryRun
-      ? chalk.green(`${totalFiles} files will be written`)
-      : chalk.green(`${totalFiles} files have been written successfully`);
-    const writeOutput = writeResults.aspectsWritersResults
-      .map((aspectWritersResults) => this.getOneAspectOutput(aspectWritersResults))
-      .join('\n\n');
-    return `${writeTitle}\n${writeOutput}`;
-  }
-  private getOneAspectOutput(aspectWritersResults: AspectWritersResults): string {
-    const title = `The following paths are according to aspect ${chalk.bold(aspectWritersResults.aspectId.toString())}`;
-    const writersOutput = aspectWritersResults.writersResult
-      .map((oneWritersResults) => this.getOneWriterOutput(oneWritersResults))
-      .join('\n\n');
-    return `${title}\n${writersOutput}`;
-  }
-  private getOneWriterOutput(oneWritersResults: OneConfigFileWriterResult): string {
-    const title = `The following paths are according to writer ${chalk.bold(oneWritersResults.name)}`;
-    const realConfigFilesOutput = this.getRealConfigFilesOutput(oneWritersResults.configFiles);
-    const extendingConfigFilesOutput = this.getExtendingConfigFilesOutput(oneWritersResults.extendingConfigFiles);
-    return `${title}\n${realConfigFilesOutput}\n${extendingConfigFilesOutput}`;
-  }
-
-  private getRealConfigFilesOutput(envsWrittenConfigFiles: EnvsWrittenConfigFiles): string {
-    return envsWrittenConfigFiles
-      .map((envsWrittenConfigFile) => this.getEnvGroupConfigFilesOutput(envsWrittenConfigFile))
-      .join('\n\n');
-  }
-  private getEnvGroupConfigFilesOutput(envsWrittenConfigFile: EnvsWrittenConfigFile): string {
-    const title = `The following paths are according to env(s) ${chalk.bold(envsWrittenConfigFile.envIds.join(', '))}`;
-    const filePath = envsWrittenConfigFile.configFile.filePath;
-    return `${title}\n${filePath}`;
-  }
-  private getExtendingConfigFilesOutput(extendingConfigFiles: EnvsWrittenExtendingConfigFiles): string {
-    return extendingConfigFiles
-      .map((envsWrittenExtendingConfigFile) =>
-        this.getEnvGroupExtendingConfigFilesOutput(envsWrittenExtendingConfigFile)
-      )
-      .join('\n\n');
-  }
-  private getEnvGroupExtendingConfigFilesOutput(
-    envsWrittenExtendingConfigFile: EnvsWrittenExtendingConfigFile
-  ): string {
-    const title = `The following paths are according to env(s) ${chalk.bold(
-      envsWrittenExtendingConfigFile.envIds.join(', ')
-    )}`;
-    const extendingConfigFile = envsWrittenExtendingConfigFile.extendingConfigFile;
-    const paths = extendingConfigFile.filePaths.map((p) => `  ${p} --> ${extendingConfigFile.extendingTarget}`).join('\n');
-    return `${title}\n${paths}`;
   }
 }
